+++ backend/Infrasturcture/Data/Migrations/StoreContextModelSnapshot.cs
@@ -75,7 +75,7 @@ namespace Infrasturcture.Data.Migrations
                     b.Property<DateTime>("CreatedDate")
                         .HasColumnType("TEXT");
 
-                    b.Property<decimal>("Price")
+                    b.Property<double>("Price")
                         .HasColumnType("decimal(18,2)");
 
                     b.Property<string>("ProductDescription")
